Fix bug where the SecurityGroup was associated to the wrong NetworkInterface
The SG was associated to the first NI found and nodes have two NI (one internal and one public).
This was working sometimes but not always, since the order is not consistent between the two.

Now we associate the SG to the first NI with a PublicIP found, which should work while we have one PublicIP per node.